# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nexus Chat Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}/block:
    post:
      operationId: blockUser
      tags: [users]
      summary: Block a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Blocked
  /users/{userId}/unblock:
    post:
      operationId: unblockUser
      tags: [users]
      summary: Unblock a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Unblocked

  # ─── CONVERSATIONS ───────────────────────────────────────────────────
...
  /conversations/{conversationId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{messageId}:
    get:
      operationId: getMessage
      tags: [messages]
      summary: Get a single message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    patch:
      operationId: editMessage
      tags: [messages]
      summary: Edit a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageEdit"
      responses:
        "200":
          description: Edited message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /messages/{messageId}/react:
    post:
      operationId: reactToMessage
      tags: [messages]
      summary: Add or remove an emoji reaction
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Updated reactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{messageId}/pin:
    post:
      operationId: pinMessage
      tags: [messages]
      summary: Pin a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Pinned
  /messages/{messageId}/star:
    post:
      operationId: starMessage
      tags: [messages]
      summary: Star or unstar a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Starred
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, email, displayName, status, role, createdAt]
      properties:
        id:          { type: integer }
        username:    { type: string }
        email:       { type: string }
        displayName: { type: string }
        avatarUrl:   { type: ["string", "null"] }
        bio:         { type: ["string", "null"] }
        status:      { type: string, enum: [active, suspended, banned] }
        presence:    { type: string, enum: [online, away, offline] }
        role:        { type: string, enum: [user, moderator, admin] }
        isVerified:  { type: boolean }
        lastSeenAt:  { type: ["string", "null"] }
        createdAt:   { type: string }
...
    Message:
      type: object
      required: [id, conversationId, senderId, content, type, status, createdAt]
      properties:
        id:             { type: integer }
        conversationId: { type: integer }
        senderId:       { type: integer }
        sender:
          $ref: "#/components/schemas/User"
        content:        { type: string }
        type:           { type: string, enum: [text, image, video, audio, file, sticker, gif, system] }
        status:         { type: string, enum: [sending, sent, delivered, seen] }
        replyToId:      { type: ["integer", "null"] }
        replyTo:
          $ref: "#/components/schemas/MessagePreview"
        mediaUrl:       { type: ["string", "null"] }
        mediaType:      { type: ["string", "null"] }
        mediaSize:      { type: ["integer", "null"] }
        mediaName:      { type: ["string", "null"] }
        isEdited:       { type: boolean }
        isPinned:       { type: boolean }
        isStarred:      { type: boolean }
        isDeleted:      { type: boolean }
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/Reaction"
        createdAt:      { type: string }
        editedAt:       { type: ["string", "null"] }
    MessagePreview:
      type: object
      required: [id, content, type, senderId]
      properties:
        id:         { type: integer }
        content:    { type: string }
        type:       { type: string }
        senderId:   { type: integer }
        senderName: { type: string }
    MessageInput:
      type: object
      required: [content, type]
      properties:
        content:   { type: string }
        type:      { type: string, enum: [text, image, video, audio, file, sticker, gif] }
        replyToId: { type: integer }
        mediaUrl:  { type: string }
        mediaType: { type: string }
        mediaSize: { type: integer }
        mediaName: { type: string }
    MessageEdit:
      type: object
      required: [content]
      properties:
        content: { type: string }
    ReactionInput:
      type: object
      required: [emoji]
      properties:
        emoji: { type: string }
    Reaction:
      type: object
      required: [emoji, count, userIds]
      properties:
        emoji:   { type: string }
        count:   { type: integer }
        userIds:
          type: array
          items: { type: integer }

    # ─── GROUPS ──────────────────────────────────────────────────────
...