# lib/api-zod/src/generated/types/messagePreview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus Chat Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MessagePreview {
  id: number;
  content: string;
  type: string;
  senderId: number;
  senderName?: string;
}